backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic==2.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
python-dotenv==1.0.1
bcrypt==4.1.2
python-multipart==0.0.9
